+++ src/flask_marshmallow/__init__.py
@@ -32,7 +32,15 @@ else:
 
 __version__ = "0.15.0"
 __version_info__ = Version(__version__).release
-__all__ = ["EXTENSION_NAME", "Marshmallow", "Schema", "fields", "exceptions", "pprint"]
+__all__ = [
+    "EXTENSION_NAME",
+    "Marshmallow",
+    "Schema",
+    "fields",
+    "exceptions",
+    "pprint",
+    "validate",
+]
 
 EXTENSION_NAME = "flask-marshmallow"